Updated so game can be played to any set score
A JOptionPane pops up after the game asks the users for their names. This optionPane asks the user for the win score (so the user can choose what the score limit is). 
If the user types in garbage (anything but a number) then the optionPane keeps popping up, forcing the user to type in a number. Yay :)